+++ filetools/_filetools.py
@@ -107,40 +107,7 @@ def listAllFiles(folder, **kwargs):
 	return file_list
 
 
-def searchForDuplicateFiles(folder, by = 'name'):
-	""" Searches for duplicate files in a folder.
-		Parameters
-		----------
-			folder: string [Path]
-				The folder to search through. subfolders will be included.
-			by: {'md5', 'name', 'size'}; default 'md5'
-				The method to qualify two files as being the same.
-		Return
-		------
-			duplicates: list<list<string>>
-				A list of paired filenames representing identical files.
-	"""
-	all_files = listAllFiles(folder)
 
-	checked_files = dict()
-	_duplicate_keys = list()
-	for filename in all_files:
-		if by == 'md5':
-			file_key = generateFileMd5(filename)
-		elif by == 'name':
-			file_key = os.path.basename(filename)
-		elif by == 'size':
-			file_key = os.path.getsize(filename)
-
-		if file_key in checked_files:
-			checked_files[file_key].append(filename)
-			_duplicate_keys.append(file_key)
-		else:
-			checked_files[file_key] = [filename]
-	_duplicates = list()
-	for key in _duplicate_keys:
-		_duplicates.append(checked_files[key])
-	return _duplicates
 
 def searchFiles(string, folder):
 	found = []
